machine-learning-2/tp-final/training/requirements.txt
numpy
pandas
mlflow
sklearn
imblearn
google-cloud-storage
psycopg2-binary